Address PR review: example_queue_state fixture, JobInfo module, shared log helper
- Replace the example_queue_state() helper with an example_queue_state
  pytest fixture.
- Move JobInfo into its own _job_info.py module, imported by both
  _queue_state.py and _queue_utils.py.
- Move the issues-test log-scaffolding helper into testing_utilities as
  write_attempt_logs.
