+++ tests/dandi_compute_code/queue/model/integration/test_process_queue_mock_queue_state.py
@@ -4,15 +4,14 @@ from unittest import mock
 import pytest
 
 from dandi_compute_code.queue import QueueState
-from model.testing_utilities import example_queue_state
 
 
 @pytest.mark.ai_generated
 def test_process_queue_skips_refresh_when_state_non_empty(
-    queue_directory: pathlib.Path, processing_directory: pathlib.Path
+    example_queue_state: QueueState, queue_directory: pathlib.Path, processing_directory: pathlib.Path
 ) -> None:
     """process_queue runs without warning when state.jsonl already has entries."""
-    example_queue_state().to_file(queue_directory / "state.jsonl")
+    example_queue_state.to_file(queue_directory / "state.jsonl")
 
     with (
         mock.patch(
@@ -29,10 +28,10 @@ def test_process_queue_skips_refresh_when_state_non_empty(
 
 @pytest.mark.ai_generated
 def test_process_queue_submits_when_no_jobs_running(
-    queue_directory: pathlib.Path, processing_directory: pathlib.Path
+    example_queue_state: QueueState, queue_directory: pathlib.Path, processing_directory: pathlib.Path
 ) -> None:
     """process_queue requests two submissions when no AIND jobs are running."""
-    example_queue_state().to_file(queue_directory / "state.jsonl")
+    example_queue_state.to_file(queue_directory / "state.jsonl")
 
     with (
         mock.patch(
@@ -49,10 +48,10 @@ def test_process_queue_submits_when_no_jobs_running(
 
 @pytest.mark.ai_generated
 def test_process_queue_respects_explicit_max_concurrent_jobs(
-    queue_directory: pathlib.Path, processing_directory: pathlib.Path
+    example_queue_state: QueueState, queue_directory: pathlib.Path, processing_directory: pathlib.Path
 ) -> None:
     """process_queue uses the explicit AIND concurrency limit to compute submissions."""
-    example_queue_state().to_file(queue_directory / "state.jsonl")
+    example_queue_state.to_file(queue_directory / "state.jsonl")
 
     with (
         mock.patch(
@@ -72,10 +71,10 @@ def test_process_queue_respects_explicit_max_concurrent_jobs(
 
 @pytest.mark.ai_generated
 def test_process_queue_does_not_submit_when_jobs_running(
-    queue_directory: pathlib.Path, processing_directory: pathlib.Path
+    example_queue_state: QueueState, queue_directory: pathlib.Path, processing_directory: pathlib.Path
 ) -> None:
     """process_queue does not submit when two AIND-Ephys-Pipeline jobs already run."""
-    example_queue_state().to_file(queue_directory / "state.jsonl")
+    example_queue_state.to_file(queue_directory / "state.jsonl")
 
     with (
         mock.patch(
@@ -92,10 +91,10 @@ def test_process_queue_does_not_submit_when_jobs_running(
 
 @pytest.mark.ai_generated
 def test_process_queue_submits_one_when_one_job_running(
-    queue_directory: pathlib.Path, processing_directory: pathlib.Path
+    example_queue_state: QueueState, queue_directory: pathlib.Path, processing_directory: pathlib.Path
 ) -> None:
     """process_queue requests one submission when exactly one AIND-Ephys-Pipeline job is running."""
-    example_queue_state().to_file(queue_directory / "state.jsonl")
+    example_queue_state.to_file(queue_directory / "state.jsonl")
 
     with (
         mock.patch(
@@ -112,10 +111,10 @@ def test_process_queue_submits_one_when_one_job_running(
 
 @pytest.mark.ai_generated
 def test_process_queue_passes_processing_directory_to_submit_next(
-    queue_directory: pathlib.Path, processing_directory: pathlib.Path
+    example_queue_state: QueueState, queue_directory: pathlib.Path, processing_directory: pathlib.Path
 ) -> None:
     """process_queue forwards processing_directory to _submit_next when idle."""
-    example_queue_state().to_file(queue_directory / "state.jsonl")
+    example_queue_state.to_file(queue_directory / "state.jsonl")
 
     with (
         mock.patch(
@@ -131,9 +130,11 @@ def test_process_queue_passes_processing_directory_to_submit_next(
 
 
 @pytest.mark.ai_generated
-def test_process_queue_forwards_test_flag(queue_directory: pathlib.Path, processing_directory: pathlib.Path) -> None:
+def test_process_queue_forwards_test_flag(
+    example_queue_state: QueueState, queue_directory: pathlib.Path, processing_directory: pathlib.Path
+) -> None:
     """process_queue forwards test=True to _submit_next."""
-    example_queue_state().to_file(queue_directory / "state.jsonl")
+    example_queue_state.to_file(queue_directory / "state.jsonl")
 
     with (
         mock.patch(
